pyproject.toml
[project]
name = "autoshift"
version = "1.1.0"
description = "AutoSHIFt: Automatically redeem Gearbox SHIFt Codes"
authors = [
    { name = "Fabian Schweinfurth", email = "fabbi@users.noreply.github.com"}
]
license = "MIT"

requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.5",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "peewee>=3.18.2",
    "pydantic-settings>=2.10.1",
    "rich>=14.1.0",
    "typer-slim>=0.17.4",
]

[project.scripts]
auto = "src:auto.app"

[dependency-groups]
dev = [
    "debugpy>=1.8.16",
    "ipdb>=0.13.13",
    "ruff>=0.13.0",
    "types-peewee>=3.18.2.20250710",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
src = { path = "./src" }

[tool.pyright]
typeCheckingMode = "standard"
ignore = ["typings/**", "**/**.pyi"]
include = ["src"]
reportUnusedImport = "warning"
reportMissingTypeStubs = false                                              #"information"
reportUnknownMemberType = false
reportMissingParameterType = false                                          #"hint"
reportUnknownParameterType = false
reportExplicitAny = false
reportAny = false
reportUnannotatedClassAttribute = false
reportUnusedCallResult = false
reportPrivateImportUsage = false
reportIncompatibleVariableOverride = "warning"

[tool.ruff]
line-length = 90
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "static",
    "typings",
    ".vscode",
    "lib",
]
builtins = ["reveal_type"]

[tool.ruff.lint]
select = [
    "E4", "E7", "E9", "F", ### DEFAULT
    "UP",
    "TID252",
    "I",
    "TD005",  # todo description
    "TD006",
    "TD007",
    "FIX",
    "F811",   # redefinition of unused
    "ASYNC",
    # code quality
    "RUF100", # unused `noqa` directive
    "RUF051", # if-key-in-dict-del
    # "RUF029", # unused async             ### PREVIEW feature
    # "RUF056", # falsy-dict-get-fallback  ### PREVIEW feature
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
